requirements.txt
pandas 
numpy 
seaborn 
streamlit 
joblib
plotly.express
plotly.subplots
plotly.graph_objects
os 
pyautogui
time